More testing. Made Float the most generic number type

+++ profiler/src/main/scala/com/scottlogic/deg/classifier/SemanticType.scala
@@ -3,9 +3,9 @@ package com.scottlogic.deg.classifier
 sealed trait SemanticType {def rank: Int}
 case object NullType extends SemanticType {val rank = 0}
 case object StringType extends SemanticType {val rank = 1}
-case object IntegerType extends SemanticType {val rank = 2}
+case object FloatType extends SemanticType {val rank = 2}
 case object DoubleType extends SemanticType {val rank = 3}
-case object FloatType extends SemanticType {val rank = 4}
+case object IntegerType extends SemanticType {val rank = 4}
 case object EmailType extends SemanticType {val rank = 5}
 case object NameType extends SemanticType {val rank = 6}
 case object CountryCodeType extends SemanticType {val rank = 7}
